+++ build.py
@@ -1,13 +1,21 @@
 #!/usr/bin/env python3
 import sbuildr
+import sbuildr.dependencies.builders as builders
+import sbuildr.dependencies.fetchers as fetchers
+
 import glob
+import os
+
+stdlib = sbuildr.Library("stdc++")
 project = sbuildr.Project()
 
-libstest = project.library("stest", sources=["STest.cpp"], libs=["stdc++"])
-project.install(libstest, "/usr/local/lib")
-project.install("STest.hpp", "/usr/local/include/Stealth")
+slog = sbuildr.dependencies.Dependency(fetchers.GitFetcher("https://github.com/pmarathe25/SLog", tag="v0.2.1"), builders.SBuildrBuilder())
+libstest = project.library("stest", sources=["STest.cpp"], libs=[stdlib])
+
+# Need stream overloads from the SLog header
+project.interfaces(["STest.hpp"], depends=[slog])
 
 for source in glob.iglob("tests/*.cpp"):
-    project.test(source.split('.')[0], sources=[source], libs=["stdc++", libstest])
+    project.test(os.path.basename(source).split('.')[0], sources=[source], libs=[stdlib, libstest])
 
-sbuildr.cli(project, default_profiles=["debug"])
+project.export()